Refactor code placement
